refactor(@angular-devkit/build-angular): move internal bundle output processing into esbuild bundle helper
The path adjustments and initial file logic for index file generation are now centralized in the esbuild
build helper function. This allows usage of the functionality in multiple areas of the build system.
